+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - AssetsService.GetAssetParams
 - new proto файлы (Версия 2.4.0 (05.06.2025))
 - client.NewAccountRequest(accountId string).Do(ctx) = обертка для более "простого" получения данных по счету  
-
+- client.GetTime(ctx context.Context) (time.Time, error)
+- 
 ## [0.1.0] - 2025-04-17
 ### Added
 - реализован (почти) весь функционал для работы с api.